Commonize datamodule providers where possible for reuse
